==> notebooks/cellchat_benchmark/01_download_atlas.ipynb <==
import os

from googleapiclient.http import MediaIoBaseDownload
from googleapiclient.discovery import build
from apiclient import discovery
from httplib2 import Http
import oauth2client
from oauth2client import file, client, tools


data_path = '/data2/hratch/immune_CCI/covid/covid_atlas/'
root_path = '/home/hratch/Projects/c2c_sim/cellchat_for_tensor/'

# https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE158055

hls = ['https://ftp.ncbi.nlm.nih.gov/geo/series/GSE158nnn/GSE158055/suppl/GSE158055%5Fcovid19%5Fcounts%2Emtx%2Egz',
       'https://ftp.ncbi.nlm.nih.gov/geo/series/GSE158nnn/GSE158055/suppl/GSE158055%5Fcell%5Fannotation%2Ecsv%2Egz',
       'https://ftp.ncbi.nlm.nih.gov/geo/series/GSE158nnn/GSE158055/suppl/GSE158055%5Fsample%5Fmetadata%2Exlsx',
       'https://ftp.ncbi.nlm.nih.gov/geo/series/GSE158nnn/GSE158055/suppl/GSE158055%5Fcovid19%5Fbarcodes%2Etsv%2Egz',
       'https://ftp.ncbi.nlm.nih.gov/geo/series/GSE158nnn/GSE158055/suppl/GSE158055%5Fcovid19%5Ffeatures%2Etsv%2Egz'
      ]
for hl in hls:
    cmd = 'wget -P ' + data_path + 'raw/ ' + hl 
    os.system(cmd)

# https://stackoverflow.com/questions/37453841/download-a-file-from-google-drive-using-wget
hl = 'https://drive.google.com/u/2/open?id=1IwWcn4W-YKgNbz4DpNweM2cKxlx1hbM0'
file_id = '1IwWcn4W-YKgNbz4DpNweM2cKxlx1hbM0'
# git clone https://github.com/chentinghao/download_google_drive.git@e1097f50e631043f8f0d5d23b0fee2b6cb3eaa9b
downloader = '/data2/hratch/Software/download_google_drive/' 
cmd = 'python ' + downloader + 'download_gdrive.py ' + file_id + ' ' + data_path + 'raw/COVID19_ALL.h5ad.tar.gz'
os.system(cmd)

# unzip
cmd = 'tar -xvzf ' + data_path + 'raw/COVID19_ALL.h5ad.tar.gz ' +  '-C ' + data_path + 'raw/'
os.system(cmd)
