==> notebooks/cellchat_benchmark/03_cellchat_preprocess_timing.ipynb <==
suppressPackageStartupMessages({
    library(CellChat)
    library(patchwork)
    library(RhpcBLASctl)
    library(Matrix)
})
options(stringsAsFactors = FALSE)
input_path_timing = '/data2/hratch/immune_CCI/covid/covid_atlas/interim/timing_inputs/'

md.cell <- read.csv(paste0(input_path_timing,'metadata_for_timing.csv'), row.names = 1)

sample_names<-unique(md.cell$sampleID)
sample_context_map<-list()
for (sn in sample_names){
    context<-unique(md.cell[md.cell$sampleID == sn,'CoVID.19.severity'])
    if (length(context) != 1){stop('Incorred sample to context mapping')}
    sample_context_map[[sn]]<-context
}
saveRDS(sample_context_map, paste0(input_path_timing, 'sample_context_map.rds'))
